% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpexrand.R
\name{print.metaan.ara}
\alias{print.metaan.ara}
\title{Pooled excess risk estimate using the alternative random effect model meta-analysis}
\usage{
\method{print}{metaan.ara}(x, ...)
}
\arguments{
\item{x}{Object of class metaan.ara}

\item{...}{Other arguments}
}
\value{

}
\description{
Alternative fixed effect model for alternative meta-analysis of excess relative risk (ERR) or excess odds ratio (EOR) estimates.
}
