% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpexfix.R
\name{print.metaan.arf}
\alias{print.metaan.arf}
\title{Pooled excess risk estimate using the alternative fixed effect model meta-analysis}
\usage{
\method{print}{metaan.arf}(x, ...)
}
\arguments{
\item{x}{Object of class metaan.arf}

\item{...}{Other arguments}
}
\value{

}
\description{
Alternative fixed effect model for meta-analysis of excess relative risk (ERR) or excess odds ratio (EOR) estimates.
}
