% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexfix.R
\name{print.metaan.era}
\alias{print.metaan.era}
\title{Pooled excess risk estimate using the fixed effect model meta-analysis}
\usage{
\method{print}{metaan.era}(x, ...)
}
\arguments{
\item{x}{Object of class metaan.era}

\item{...}{Other arguments}
}
\value{

}
\description{
Fixed effect model for standard meta-analysis of excess relative risk (ERR) or excess odds ratio (EOR) estimates.
}
