% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexrand.R
\name{print.metaan.erf}
\alias{print.metaan.erf}
\title{Pooled excess risk estimate using the random effect model meta-analysis}
\usage{
\method{print}{metaan.erf}(x, ...)
}
\arguments{
\item{x}{Object of class metaan.erf}

\item{...}{Other arguments}
}
\value{

}
\description{
Random effect model for standard meta-analysis of excess relative risk (ERR) or excess odds ratio (EOR) estimates.
}
