% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estmeta.R
\name{print.metaan.est}
\alias{print.metaan.est}
\title{Meta-analysis of Beta (the parameters or coefficients estimated) from regression models}
\usage{
\method{print}{metaan.est}(x, ...)
}
\arguments{
\item{x}{Object of class metaan.est}

\item{...}{Other arguments}
}
\value{

}
\description{
Fixed effect model or DerSimonian and Laird-based Random effect model for standard meta-analysis of Beta (the parameters or coefficients) estimated from regression models (e.g linear regression or generalised linear regression models).
}
