% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priskran.R
\name{print.metaan.ra}
\alias{print.metaan.ra}
\title{Pooled risk estimate using the random effect model meta-analysis}
\usage{
\method{print}{metaan.ra}(x, ...)
}
\arguments{
\item{x}{Object of class metaan.ra}

\item{...}{Other arguments}
}
\value{

}
\description{
Random effect model for standard meta-analysis of risk estimate (e.g relative risk (RR), odds ratio (OR) and hazard ratio (HR))
}
