% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priskfix.R
\name{print.metaan.rr}
\alias{print.metaan.rr}
\title{Pooled risk estimate using the fixed effect model meta-analysis}
\usage{
\method{print}{metaan.rr}(x, ...)
}
\arguments{
\item{x}{Object of class metaan.rr}

\item{...}{Other arguments}
}
\value{

}
\description{
Fixed effect model for standard meta-analysis of risk estimate (e.g relative risk (RR), odds ratio (OR) or hazard ratio (HR))
}
