% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexfix.R
\name{pexfix}
\alias{pexfix}
\title{Pooled excess risk estimate using the fixed effect model meta-analysis}
\usage{
pexfix(err, u, l, type = "excess", test = "FIXE")
}
\arguments{
\item{err}{A numeric vector of the risk estimated from the individual studies}

\item{u}{A numeric vector of the upper bound of the confidence interval of the risk estimated from the individual studies.}

\item{l}{A numeric vector of the lower bound of the confidence interval of the risk estimated from the individual studies.}

\item{type}{Logical indicating the method to be used. The default is "excess" indicating that risk estimate model should be used.}

\item{test}{Logical indicating the statistical method to be used. The default if FIXE for the fixe effect model.}
}
\description{
Fixed effect model for standard meta-analysis of excess relative risk (ERR) or excess odds ratio (EOR) estimates.
}
\examples{
study <- c("Canada", "Northern USA", "Chicago", "Georgia","Puerto", "Comm",
"Madanapalle", "UK", "South Africa", "Haiti", "Madras")
Risk <- c(0.205, 0.411, 0.254, 1.562, 0.712, 0.983, 0.804, 0.237, 0.625,
0.198, 1.012)
lower_ci <- c(0.086, 0.134, 0.149, 0.374, 0.573, 0.582, 0.516, 0.179,
0.393, 0.078, 0.895)
upper_ci <- c(0.486, 1.257, 0.431, 6.528, 0.886, 1.659, 1.254, 0.312,
0.996, 0.499, 1.145)

donne <- data.frame(cbind(study, Risk, lower_ci, upper_ci))

donne$Risk <- as.numeric(as.character(donne$Risk))
donne$upper_ci <- as.numeric(as.character(donne$upper_ci))
donne$lower_ci <- as.numeric(as.character(donne$lower_ci))

pexfix(err=donne$Risk, u=donne$upper_ci, l=donne$lower_ci,
type = "excess", test = "FIXE")


}
