% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risksens.R
\name{risksens}
\alias{risksens}
\title{Sensitivity analysis for relative risk (or odds ratio) meta-analysis}
\usage{
risksens(
  study,
  rr,
  u,
  l,
  form = c("Log", "nonLog"),
  type = "risk",
  test = c("FIXED", "RANDOM"),
  conf.level = 0.95
)
}
\arguments{
\item{study}{A vector (or column for dataframe, matrix) specifying the column reporting the author's name or the individual study's name}

\item{rr}{A numeric vector of the risk estimated from the individual studies}

\item{u}{A numeric vector of the upper bound of the confidence interval of the risk estimated from the individual studies.}

\item{l}{A numeric vector of the lower bound of the confidence interval of the risk estimated from the individual studies.}

\item{form}{Logical, indicating the scale of the data. If Log, then the original data are in logarithmic scale.}

\item{type}{Logical, indicating the method to be used. The default is risk indicating that risk estimate model should be used.}

\item{test}{Logical, indicating the statistical method to be used. The user have the choice between "FIXED" for the fixed effect model, and "RANDOM" for the random effect model.}

\item{conf.level}{Coverage for confidence interval}
}
\value{
Object of class "data.frame" that print the output from the risksens function. The following could be found from the output :
\itemize{
\item Study: Indication of the study removed from the pooled effect estimate
\item Effect: The pooled effect from the individual studies' estimate (RR, OR, or HR)
\item SE-Log(Effect): The standard error of the pooled effect (see reference Richardson et al 2020 for more details)
\item Lower CI: The lower confidence interval bound of the pooled effect
\item Upper CI: The upper confidence interval bound of the pooled effect
}
}
\description{
Fixed or Random effect model could be used for the sensitivity analysis computation. The
risk estimate could be e.g relative risk (RR), odds ratio (OR) or hazard ratio (HR).

In the sensitivity analysis, each individual study is removed one at a time and the
summarized estimate is computed to access the effect of the removed study on the overall pooled estimate.
}
\examples{
study <- c("Canada", "Northern USA", "Chicago", "Georgia","Puerto", "Comm", "Madanapalle",
"UK", "South Africa", "Haiti", "Madras")
Risk <- c(0.205, 0.411, 0.254, 1.562, 0.712, 0.983, 0.804, 0.237, 0.625, 0.198, 1.012)
lower_ci <- c(0.086, 0.134, 0.149, 0.374, 0.573, 0.582, 0.516, 0.179, 0.393, 0.078, 0.895)
upper_ci <- c(0.486, 1.257, 0.431, 6.528, 0.886, 1.659, 1.254, 0.312, 0.996,
0.499, 1.145)

donne <- data.frame(cbind(study, Risk, lower_ci, upper_ci))

donne$Risk <- as.numeric(as.character(donne$Risk))
donne$upper_ci <- as.numeric(as.character(donne$upper_ci))
donne$lower_ci <- as.numeric(as.character(donne$lower_ci))

# on the log form
donne$ln_risk <- log(donne$Risk)
donne$ln_lower_ci <- log(donne$lower_ci)
donne$ln_upper_ci <- log(donne$upper_ci)

risksens(study=donne$study, rr=donne$ln_risk, u=donne$ln_upper_ci, l=donne$ln_lower_ci,
type="risk", form="Log", test = "FIXED")

risksens(study=donne$study, rr=donne$ln_risk, u=donne$ln_upper_ci, l=donne$ln_lower_ci,
type="risk", form="Log", test = "RANDOM")

risksens(study=donne$study, rr=donne$Risk, u=donne$upper_ci, l=donne$lower_ci,
type="risk", form="nonLog", test = "FIXED")

risksens(study=donne$study, rr=donne$Risk, u=donne$upper_ci, l=donne$lower_ci,
type="risk", form="nonLog", test = "RANDOM")

}
\author{
Kossi Abalo
}
